confluence/requirements.txt
atlassian-python-api==3.28.0
beautifulsoup4==4.11.1
click==8.1.3
python-dotenv==0.20.0
lxml==4.9.1
pylint==2.4.4
pytest==5.3.5
